refactor(utils): remove setCypressElementsByButtonText
