+++ src/assertions/button.ts
@@ -1,6 +1,10 @@
 import { Then } from '@badeball/cypress-cucumber-preprocessor';
 
-import { setCypressElementsByButtonText } from '../utils';
+import {
+  When_I_find_button_by_text,
+  When_I_find_buttons_by_text,
+} from '../queries';
+import { getCypressElement } from '../utils';
 
 /**
  * Then I see button:
@@ -22,7 +26,8 @@ import { setCypressElementsByButtonText } from '../utils';
  * - {@link Then_I_see_text | Then I see text}
  */
 export function Then_I_see_button(text: string) {
-  setCypressElementsByButtonText(text).should('exist');
+  When_I_find_button_by_text(text);
+  getCypressElement().should('exist');
 }
 
 Then('I see button {string}', Then_I_see_button);
@@ -47,7 +52,8 @@ Then('I see button {string}', Then_I_see_button);
  * - {@link Then_I_do_not_see_text | Then I do not see text}
  */
 export function Then_I_do_not_see_button(text: string) {
-  setCypressElementsByButtonText(text).should('not.exist');
+  When_I_find_buttons_by_text(text);
+  getCypressElement().should('not.exist');
 }
 
 Then('I do not see button {string}', Then_I_do_not_see_button);
