+++ src/queries/button.ts
@@ -1,9 +1,6 @@
 import { When } from '@badeball/cypress-cucumber-preprocessor';
 
-import {
-  setCypressElementByButtonText,
-  setCypressElementsByButtonText,
-} from '../utils';
+import { getCypressElement, setCypressElement } from '../utils';
 
 /**
  * When I find buttons by text:
@@ -33,7 +30,22 @@ import {
  * - {@link When_I_find_button_by_text | When I find button by text}
  */
 export function When_I_find_buttons_by_text(text: string) {
-  setCypressElementsByButtonText(text);
+  const buttons = [
+    'button',
+    "[type='button']",
+    "[type='submit']",
+    "[role='button']",
+  ];
+
+  const selector1 = buttons
+    .map((selector) => `${selector}:contains(${JSON.stringify(text)}):visible`)
+    .join(',');
+
+  const selector2 = buttons
+    .map((selector) => `${selector}[value=${JSON.stringify(text)}]:visible`)
+    .join(',');
+
+  setCypressElement(cy.get(`${selector1},${selector2}`));
 }
 
 When('I find buttons by text {string}', When_I_find_buttons_by_text);
@@ -67,7 +79,8 @@ When('I find buttons by text {string}', When_I_find_buttons_by_text);
  * - {@link When_I_find_buttons_by_text | When I find buttons by text}
  */
 export function When_I_find_button_by_text(text: string) {
-  setCypressElementByButtonText(text);
+  When_I_find_buttons_by_text(text);
+  setCypressElement(getCypressElement().first());
 }
 
 When('I find button by text {string}', When_I_find_button_by_text);
