+++ src/utils/index.ts
@@ -1,4 +1,3 @@
-export * from './button';
 export * from './display-value';
 export * from './element';
 export * from './label';
